+++ src/main/java/roomescape/domain/reservation/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -11,9 +11,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.repository.ReservationRepository;
+import roomescape.domain.reservation.dto.ReservationRequestDTO;
+import roomescape.domain.reservation.dto.ReservationResponseDTO;
 import roomescape.domain.time.ReservationTime;
 import roomescape.domain.user.User;
-import roomescape.domain.time.ReservationTimeRepository;
+import roomescape.domain.time.repository.ReservationTimeRepository;
 
 @Controller
 public class ReservationController {
