+++ src/main/java/roomescape/domain/reservation/repository/ReservationRepository.java
@@ -1,6 +1,7 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.repository;
 
 import java.util.List;
+import roomescape.domain.reservation.Reservation;
 
 public interface ReservationRepository {
     // 예약 추가
