+++ src/test/java/roomescape/LayerSeparationTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.reservation.ReservationController;
+import roomescape.domain.reservation.controller.ReservationController;
 
 @SpringBootTest
 public class LayerSeparationTest {
